Revert "Implement Carousel using Bootstrap"
