+++ package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "bootstrap": "^5.3.1",
     "react": "^18.2.0",
-    "react-bootstrap": "^2.8.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.14.1"
   },
