--- src/components/ProjectsCarousel/ProjectsCarousel.css
@@ -1,3 +1,0 @@
-.project-image {
-  width: 60%;
-}
