--- src/components/ReturnButton/ReturnButton.jsx
@@ -1,14 +1,0 @@
-import { Link } from 'react-router-dom';
-
-const ReturnButton = () => {
-  return (
-    <div>
-      <Link
-        className='home-link'
-        to='/'
-      />
-    </div>
-  )
-}
-
-export default ReturnButton;
