+++ src/pages/Certificates/Certificates.jsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { fetchAPI } from '../../helpers/fetchAPI';
 import { certificates_URL } from '../../helpers/info';
 import './Certificates.css';
-import ReturnButton from '../../components/ReturnButton/ReturnButton';
+import { Link } from 'react-router-dom';
 
 const Certificates = () => {
   const [certificates, setCertificates] = useState([]);
@@ -18,8 +18,11 @@ const Certificates = () => {
   }, []);
 
   return (
-    <div className='certificates-container text-center p-3 h-100 mx-auto'>
-      <ReturnButton />
+    <article className='certificates-container text-center p-3 h-100 mx-auto'>
+      <Link
+        className='home-link'
+        to='/'
+      />
       <h2>Certificates</h2>
       <section className='d-sm-flex flex-column d-md-grid'>
         <div className='certificates-container row g-md-1 justify-content-center' style={{ gap: '1rem' }}>
@@ -41,7 +44,7 @@ const Certificates = () => {
         ))}
         </div>
       </section>
-    </div>
+    </article>
   )
 }
