+++ src/pages/Projects/Projects.jsx
@@ -1,14 +1,34 @@
+import { useState, useEffect } from 'react';
+import { repos_URL } from '../../helpers/info';
+import { fetchAPI } from '../../helpers/fetchAPI';
+import Card from '../../components/Card/Card';
+import { Link } from 'react-router-dom';
 import './Projects.css';
-import ProjectsCarousel from '../../components/ProjectsCarousel/ProjectsCarousel';
-import ReturnButton from '../../components/ReturnButton/ReturnButton';
 
 const Projects = () => {
+  const [projects, setProjects] = useState([]);
+
+  useEffect(() => {
+    const fetchProjects = async () => {
+      const response = await fetchAPI(repos_URL);
+      const filteredProjects = response.filter((project) => !project.name.includes('biancashiromoto'));
+      const sortedProjects = filteredProjects.sort((a, b) => new Date(b.created_at) - new Date(a.created_at));
+      setProjects(sortedProjects);
+    }
+    fetchProjects();
+  }, []);
+
   return (
-    <div className='projects-container text-center p-3 mx-auto vh-100'>
-      <ReturnButton />
+    <article className='projects-container text-center p-3 vh-100'>
+      <Link
+        className='home-link'
+        to='/'
+      />
       <h2>Projects</h2>
-      <ProjectsCarousel />
-    </div>
+      {projects.map((project) => (
+        <Card key={project.id} props={ project } />
+      ))}
+    </article>
   )
 }
